reorganize layout components
